fixed capatilization and white space bugs

+++ routes.py
@@ -4,7 +4,7 @@ from flask import Flask, render_template, request, redirect, flash
 from flask_sqlalchemy import SQLAlchemy
 from wtforms import StringField, validators, SubmitField, ValidationError
 from flask_wtf import FlaskForm
-import forms
+from forms import Add, Delete
 
 # declaring database location
 project_dir = os.path.dirname(os.path.abspath(__file__))
@@ -28,10 +28,10 @@ db.create_all()
 # Home page route
 @app.route("/", methods=["GET", "POST"])
 def home():
-    form_add = forms.Add() # Defining forms
-    form_delete = forms.Delete()
-    if form_add.validate_on_submit(): # Adding item into database,
-        item_name = form_add.item.data.strip()
+    form_add = Add() # Defining forms
+    form_delete = Delete()
+    if request.method == 'POST' and form_add.validate_on_submit(): # Adding item into database,
+        item_name = " ".join(form_add.item.data.lower().split()) # Gets rid of extra spaces before, inbetween, and after text
         item = Backpack(item=item_name, amount=1)
         database_item = Backpack.query.filter_by(item=item_name).first()
         if database_item == None: # checking if item is in database
@@ -48,40 +48,48 @@ def home():
 # Route for delete function
 @app.route('/delete', methods=["POST", "GET"])
 def delete():
-    form = forms.Delete() # Defining delete form
+    form = Delete() # Defining delete form
     if request.method == 'POST' and form.validate():
-        item_name = form.item.data.strip()
+        item_name = " ".join(form.item.data.lower().split()) # Gets rid of extra spaces before, inbetween, and after text
         backpack_item = Backpack.query.filter_by(item=item_name).first()
-        if backpack_item == None:
+        if backpack_item == None: # Checks if the item is in the database
             print("no such item in db")
-        elif backpack_item.amount > 1:
-            backpack_item.amount = backpack_item.amount - 1
-            db.session.commit()
         else:
             db.session.delete(backpack_item)
+            flash("Successfully Deleted {}".format(backpack_item.item))
             db.session.commit()
     return redirect('/')
 
+# Route for incrementing amounts of an item in the backpack
 @app.route('/increase', methods=["POST", "GET"])
 def increase():
     if request.method == 'POST':
         item = request.form.get("item_name")
-        database_item = Backpack.query.filter_by(item=item).first()
-        database_item.amount = database_item.amount + 1
-        db.session.commit()
+        if item != None:
+            item = item.strip().lower()
+            database_item = Backpack.query.filter_by(item=item).first() # Searches database for item and increments it
+            database_item.amount = database_item.amount + 1 
+            db.session.commit()
+        else:
+            flash("Item was not in database")
     return redirect('/')
 
+# Route for decreasing the amount of an item in the backpack
 @app.route('/decrease', methods=["POST", "GET"])
 def decrease():
     if request.method == 'POST':
         item = request.form.get("item_name")
-        database_item = Backpack.query.filter_by(item=item).first()
-        if database_item.amount > 1:
-            database_item.amount = database_item.amount - 1
-            db.session.commit()
+        if item != None:
+            item = item.strip().lower()
+            database_item = Backpack.query.filter_by(item=item).first() # Searches database for item and increments it
+            if database_item.amount > 1:
+                database_item.amount = database_item.amount - 1
+                db.session.commit()
+            else:
+                db.session.delete(database_item)
+                db.session.commit()
         else:
-            db.session.delete(database_item)
-            db.session.commit()
+            flash("Item was not in database")
     return redirect('/')
 
 @app.errorhandler(404)
